requirements.txt
streamlit
biopython
numpy
py3Dmol
langchain
openai
colabfold
pandas
matplotlib
requests
